apps/api/pyproject.toml
[project]
name = "agenthive-api"
version = "0.1.0"
description = "AgentHive API"
authors = [
    {name = "Satyam Sharma",email = "satyam17102003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "pydantic[email] (>=2.11.10,<3.0.0)",
    "bcrypt (==4.0.1)",
    "pytest (>=8.4.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "fakeredis (>=2.31.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

